Add yum repositories using yum_repository resource instead of RPM
Fetching the RPM doesn't currently work for 3.4 because the version is
suffixed with -2 rather than -1.

If the yum_repository resource was not used before to avoid the
additional cookbook dependency, this is no longer an issue.

I wasn't sure whether to add the non-supported repository but I felt
it best not to modify the existing behaviour.