Simplify installation of web packages and support Debian
On RHEL, the dependencies are all correctly stated, you just need to
install zabbix-web-pgsql or zabbix-web-mysql. This was previously
hardcoded to zabbix-web-pgsql.

On Debian and Ubuntu, some dependencies do still need to be installed
manually. The recipe has been rejigged to be simpler while supporting
Debian as well as Ubuntu.