Handle SELinux in the cookbook itself and not just the tests
This may as well work out of the box!

There is no need to make this conditional on RHEL as I added a
use_selinux helper to selinux_policy a while back.

The setrlimit policies have been added to Fedora recently but only the
agent version is present in RHEL 7. This is no harm in defining it
twice in any case.

The selinux_policy_port resources should be removed if my unix_socket
branch is merged.